Revert "add AUTHORS and ChangeLog files"
AUTHORS and ChangeLog are generated by pbr every time you install
the software or make an sdist tarball. The idea is that git is
the source of truth, and the only time these files are needed is
in in a source tarball when git isn't available. Committing them
will mean you'll have new local content on every commit, which
will likely get ... annoying.

This reverts commit 30312b1260f996ddf3ea771ae8bc89483a66dd08.